scripts/requirements.txt
numpy
librosa
Pillow
scikit-learn
joblib
opencv-python
pandas